requirements.txt
Flask==3.1.3
gunicorn==21.2.0
numpy==2.2.6
openpyxl==3.1.5
pandas==2.3.3
python-docx==1.2.0
python-dotenv==1.2.2
